pyproject.toml
[project]
name = "certificate-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-pubsub>=2.34.0",
    "python-dotenv>=1.2.1",
]
